Moved Tile.js location in file structure
